src/requirements.txt
-i https://pypi.org/simple
alembic==1.13.1; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
async-timeout==4.0.3; python_full_version < '3.12.0'
asyncpg==0.29.0; python_full_version >= '3.8.0'
bcrypt==4.1.2
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
cryptography==41.0.7
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.109.0; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.4; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
python-jose[cryptography]==3.3.0
python-multipart==0.0.6; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.25; python_version >= '3.7'
starlette==0.35.1; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
uvicorn[standart]==0.26.0; python_version >= '3.8'
